% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_randomization_matrix.R
\name{create_randomization_matrix}
\alias{create_randomization_matrix}
\title{Create Randomization Matrix}
\usage{
create_randomization_matrix(units, n, block = NULL)
}
\arguments{
\item{units}{Number of units in dataset}

\item{n}{Number of permutations}

\item{block}{Numeric vector with length equal to "units"
indicating block assignments for each unit (optional)}
}
\value{
Matrix with unique randomized permutations
}
\description{
Creates randomization matrix of assignments for given number of units and
permutations.Returns matrix with unique randomized permutations.
}
\details{
Note, if the number of specified permutations exceeds the maximum number
of unique permutations, the matrix returned will contain the
maximum number of permutations.
}
\examples{
create_randomization_matrix(14,128,rep(1:7, each = 2))
}
