% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_matrix_1D.R
\docType{data}
\name{rand_matrix_1D}
\alias{rand_matrix_1D}
\title{Randomization matrix of body camera assignments for District1D}
\format{
A matrix with 10000 columns (body camera assignment options):
\describe{
\item{1}{Body Camera On}
\item{0}{Body Camera Off}
}
}
\usage{
rand_matrix_1D
}
\description{
This matrix contains 10,000 possible body camera assignments
for District1D.
0 means no camera. 1 means camera.
}
\keyword{datasets}
