% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_matrix_1DS.R
\docType{data}
\name{rand_matrix_1DS}
\alias{rand_matrix_1DS}
\title{Randomization matrix of body camera assignments for District1DS}
\format{
A matrix with 128 columns (body camera assignment options):
\describe{
\item{1}{Body Camera On}
\item{0}{Body Camera Off}
}
}
\usage{
rand_matrix_1DS
}
\description{
This matrix contains all possible body camera assignments for
District1DS.
0 means no camera. 1 means camera.
}
\keyword{datasets}
