% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Sample Data}
\format{
A table with fish treatment assignments and turn angles:
\describe{
\item{w}{Treatment Assignment: 1 indicates exposure to flashing light
and 0 indicates no exposure}
\item{turn_angle}{Angle at which the fish swim}
}
}
\usage{
sample_data
}
\description{
This dataset is for an experiment measuring the effect
of flashing lights on the movement of fish. It includes the treatment
assignments of 156 fish and the turn angles of each fish when swimming.
}
\keyword{datasets}
