% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_matrix.R
\docType{data}
\name{sample_matrix}
\alias{sample_matrix}
\title{Sample Randomization Matrix}
\format{
A matrix with 1,000 columns:
\describe{
\item{1}{Fish is Exposed to Flashing Light}
\item{0}{Fish is Not Exposed to Flashing Light}
}
}
\usage{
sample_matrix
}
\description{
This is a randomization matrix for an experiment conducted on
156 fish. This matrix contains 1,000 possible treatment assignments for
each fish.
}
\keyword{datasets}
