% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_randomization_matrix.R
\name{create_randomization_matrix}
\alias{create_randomization_matrix}
\title{Creates randomization matrix for paired data}
\usage{
create_randomization_matrix(sample_data, arrangements, units, pair)
}
\arguments{
\item{sample_data}{Sample data set. Data should have first column indicating
exposure (1) or non exposure (0) for each group (row) that is measured. Each
measured outcome (variable) should be represented by an additional column.
For paired randomization, data must be ordered so that paired units are in
subsequent rows.}

\item{arrangements}{Numbers of arrangements generated of exposure assignments
of which unique arrangements are extracted}

\item{units}{Number of units there are to measure in dataset
(One pair = control unit + experimental unit)}

\item{pair}{1 for paired randomization. 0 for non-paired randomization}
}
\value{
Matrix with unique randomized allocations for paired data
}
\description{
Returns matrix with unique randomized allocations for paired data.
}
\examples{
create_randomization_matrix(sample_district_1DS,10000,14,1)
}
